refactor(routing) move common packages
