+++ p2p/peer/queue/distance.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	peer "github.com/jbenet/go-ipfs/p2p/peer"
-	ks "github.com/jbenet/go-ipfs/routing/keyspace"
+	ks "github.com/jbenet/go-ipfs/routing/common/keyspace"
 	u "github.com/jbenet/go-ipfs/util"
 )
