+++ routing/common/keyspace/xor.go
