+++ routing/common/keyspace/xor_test.go
