+++ routing/dht/routing.go
@@ -12,8 +12,8 @@ import (
 	peer "github.com/jbenet/go-ipfs/p2p/peer"
 	"github.com/jbenet/go-ipfs/routing"
 	pb "github.com/jbenet/go-ipfs/routing/dht/pb"
-	kb "github.com/jbenet/go-ipfs/routing/kbucket"
-	record "github.com/jbenet/go-ipfs/routing/record"
+	kb "github.com/jbenet/go-ipfs/routing/common/kbucket"
+	record "github.com/jbenet/go-ipfs/routing/common/record"
 	u "github.com/jbenet/go-ipfs/util"
 	errors "github.com/jbenet/go-ipfs/util/debugerror"
 	pset "github.com/jbenet/go-ipfs/util/peerset"
